requirements.txt
python-telegram-bot==20.7
requests==2.31.0
flask==2.3.3
python-dotenv==1.0.0
aternos==2.6.0